+++ src/sub-clients/organisations/client/OrganisationsClient.ts
@@ -1,6 +1,7 @@
 import Pageable from '../../common/model/Pageable';
 import { Organisation } from '../model/Organisation';
 import { UpdateOrganisationRequest } from '../model/UpdateOrganisationRequest';
+import { User } from '../model/User';
 
 export interface OrganisationFilters {
   name?: string;
@@ -18,4 +19,6 @@ export interface OrganisationsClient {
     organisation: Organisation,
     updateAccountRequest: UpdateOrganisationRequest,
   ): Promise<Organisation>;
+
+  associateUsers(organisation: Organisation): Promise<User[]>;
 }