+++ src/sub-clients/organisations/model/UserEntity.ts
@@ -1,0 +1,6 @@
+interface UserEntity {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+}